"fixed" bandwidth calculation by adding a new function getAvgBandwidth that does what getAvgSize and getAvgSpeed do, but in one loop iteration.  Also, by calculating the bandwidth this way we can preserve a little precision.  The getAvgSpeed is still commented out and "not working", as indicated by the comments left by the previous coder.
This method relies on the precision of the timing of the
URLDownloadFile method, which I have not delved into to check its
correctness.  I did cursory testing on the overall download and found
this measurement to be adequate.  However, further exploration into
more accurate measurements should be done in the future.