changed variable with name of played_start to last_played to be more indicative of its function.
also, fixed displayed frame inconsistency with actual time.  The problem was that
it was comparing and playing against the frame->start, which I guess is supposed to
be the start of the time in which the frame is to provide a semantic summary for.
This comparison then displayed the frame at that time, which is too early!

change in actual code is:
->start ====>   ->frame_number

testing on 5 frames:
sec: frame displayed:
14:  417 ( / 30 = 13.9)
15:  428 ( / 30 = 14.2)
30:  895 ( / 30 = 29.8)
37: 1448 ( / 30 = 48.2)

so this is correct for 5 frames, and i'm assuming the rest.